photo-dedupe/requirements.txt
ImageHash==4.3.1
numpy==2.0.1
piexif==1.1.3
pillow==10.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyWavelets==1.6.0
scipy==1.14.0
